Create OneDimensional_CA_Display.ino
The following code is a simple example that demonstrates how to use an Arduino to generate a 1D cellular automaton and display it through the Serial Monitor and Serial Plotter. This code uses the Rule 30, which is known for generating chaotic patterns.

Please note that the Serial Plotter can plot numerical values over time, so for the Serial Plotter, the code will display a binary-to-decimal representation of each state in the automaton. For the Serial Monitor, it will display the states as 0's and 1's.

Upload this code to your Arduino, open the Serial Monitor to view the automaton, and then open the Serial Plotter to see the plot.

After uploading the code, open the Serial Monitor to see the 1D cellular automaton states. Then, open the Serial Plotter to see the graphical representation. Note that you cannot have both the Serial Monitor and Serial Plotter open at the same time, so you will need to switch between them.