Create OneDimensional_CA_TRNG_Seeded_Display.ino
In this refactored code, the initial state of the 1D Cellular Automaton (CA) is seeded with random values generated using the SAMD51 TRNG. It still uses Rule 30 to generate the patterns.

Upload this code to a SAMD51 board and open the Serial Monitor to see the 1D CA states. Then, open the Serial Plotter to see the graphical representation. The CA will have a random initial state due to the seed from the True Random Number Generator (TRNG) of the SAMD51.




The following code is a simple example that demonstrates how to use an Arduino to generate a 1D cellular automaton and display it through the Serial Monitor and Serial Plotter. This code uses the Rule 30, which is known for generating chaotic patterns.

Please note that the Serial Plotter can plot numerical values over time, so for the Serial Plotter, the code will display a binary-to-decimal representation of each state in the automaton. For the Serial Monitor, it will display the states as 0's and 1's.

Upload this code to your Arduino, open the Serial Monitor to view the automaton, and then open the Serial Plotter to see the plot.